Use resource service instead of inject into db directly
Implement improvements for the E2E tests and unit tests to use resource service instead of injecting directly into the database collection.

NHUB-533
