+++ e2e/cypress/support/e2e.js
@@ -60,7 +60,6 @@ export function addResources(resources) {
 export function addDefaultResources() {
     addResources([{
         resource: 'navigations',
-        use_db_collection: true,
         items: [
             NAVIGATIONS.wire.all,
             NAVIGATIONS.wire.sports,
